Added VQE
This is a notebook that is currently a WIP. It has information on how VQE works. Most of the code was obtained from the Qiskit Website (https://learning.quantum.ibm.com/tutorial/variational-quantum-eigensolver). I did explain how it works which is not provided in depth on the website. 

I will update this notebook with an example with the hydrogen atom once that is done.